app/requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-webpack-loader==1.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.27.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
oauthlib==3.2.2
openai==1.55.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
